Enumerate available slots
This introduces a ConcreteSlot which represents a RecurringSlot on a
particular day. This probably isn't the best name for it, but it works
for now.

I'm using this instead of Time object because I want to abstract slots
from time zones: the time of a slot is consistent in wall time across
summer/winter time changes.