Replace hand-built Prison fixture with factory
I was going to leave Prison.create!(...) in the before block in
`request_a_visit_spec.rb` alone.  It seemed the simplest possible
soluiton.  However, when I went to to port `visit_mailer_spec.rb` from
the old app, I realised I was going to have to either abstract or
duplicate Prison.create!(...). Neither solution seemed ideal, so I
decided it was time to go ahead and add factories using factory_girl. I
chose factory_girl as it is the most commonly used and best understood
of the fixture replacments.