WIP: add visitor mailer helpers and concerns
This is a port of most of the helpers and concerns.  It is WIP'd because
some of the shared specs are missing and so rspec won't run.  I have
left it in this state temporarily as I need to investigate a change to
the Visit#visitor/AdditionalVisitor modelling, but I don't want to lose
the work.

If this branch is still open after about 15-Dec-15 you may delete with
impunity.