Use UUID as the primary key for all objects
This came about after Paul and I discussed reference numbers in the new
system. We experimented with Base32, but realised we needed a
hexidecimal sting almost as long as a v4 UUID.  Rather than roll our own
soluiton (and curse ourselves with having to forever explain why we took
that decision), we agreed to go ahead and use UUIDs for all of the
primary keys, instead.

We discussed using UUIDs for Visits only as it is the only model that
requires them to be public facing. It was felt, however, that a
consistent approach across all tables was a wiser choice for medium and
long term development and maintenance.

NOTE: You will need to drop and re-created your database for this to
work.

CAUTION: Do NOT run this on a populated dataset you wish to keep.