Remove impossible and useless migration
Adding a non-null field with no default in one step does not work unless
the database is empty. The normal solution to this would be to perform
the operation in two steps. However, in this case, the next migration
will delete the table entirely (because we do not yet have any useful
data) so it's effectively useless.