Add detailed specs for VisitorMailer
This ensures that we have no missing localisation strings or
interpolated values, and checks expected parts of the email body. It's
much easier to find and fix errors here than in the feature specs, where
the error is hidden by the delayed mailer.

The dates in the specs are coupled to the default slots in the factory,
which isn't great, but I think it's probably no worse than laboriously
building up a graph in the spec itself.