Enhance additional visitors with JS
The additional visitors are shown when selecting the number of
additional visitors desired from the drop-down. This either shows the
hidden input elements, or hides and clears them (because we use
emptiness to determine which ones to ignore).

It's not needed without JavaScript, so the drop-down selector and
surrounding HTML is written into the document via JavaScript.