Update existing prisons when seeding
We know which prison is which thanks to the UUID. If the prison does not
exist in the database, we update it. If it does exist, we update it with
any changes.

This allows us to manipulate prisons by updated the seed data and
running db:seed on each deployment.