Validate visitor ages
In the spirit of "tell, don't ask", we delegate this to the prison,
which knows what it considers to be an adult age, and how many adults
and children are allowed (although these last two are constants at
present).

Because these validations apply to the whole step and not to an
individual field, we assign them to a 'general' error category. This,
however, means that the object must respond to 'general' in order to
avoid blowing up ActiveModel's validations.