Enable checking bookabilty by prison and date
Visitors and prison staff have different requirements for the dates they
can view.  Prison staff can see and act on all future dates, while
visitors can only work with the available slots a prison declares.

This change implements a method on the Prison model that can check if a
particluar date is included in the available slots. It is used to
indicate which days are 'Visit Days' in the visitors' booking calendar.