Rationalise I18n YAML
The difference between `key:`, `key: >`, and `key: |` can be illustrated
thus:

    a:
      one
      two
        three
    b: >
      one
      two
        three
    c: |
      one
      two
        three

results in

    a = "one two three"
    b = "one two\n  three\n"
    c = "one\ntwo\n  three\n"

In other words, we never want to use `key: >` over plain `key:`, and we
want to use `key: |` whenever it's useful to preserve the layout of
complex HTML in a localisation string.

This updates the files in `config/locales` to make it so.