Rationalise I18n YAML
There are at least nine separate ways to write a multi-line string in
YAML: https://stackoverflow.com/a/21699210

The ones that are most relevant to us can be illustrated thus:

    a:
      one
      two
    b: >
      one
      two
    c: >-
      one
      two
    d: |
      one
      two
    e: |-
      one
      two

results in

    {
      "a" => "one two",
      "b" => "one two\n",
      "c" => "one two",
      "d" => "one\ntwo\n",
      "e" => "one\ntwo",
    }

In other words, we always want to use c) in preference to b), in order
to avoid the unnecessary additional newline. The reason for choosing
form c) over a) is that c) is parsed as text, whereas a) is subject to
the usual myriad edge cases of YAML - it will blow up if it starts with
a percent sign or contains a colon, for example.

We want to use d) or e) whenever it's useful to preserve the layout of
complex HTML in a localisation string, but the absence or presence of a
final newline in this case is not important, and d) is shorter.

This updates the files in `config/locales` accordingly.