Stop prisons from processing unavailable requests
Prior to this change, the visit processing form would be shown (although
would not work) even if the visit had been withdrawn, cancelled, booked
or rejected.  This changes the prison's view such that only requested
visits show the form and can be acted upon.

I did this as part of the cancellation mailer PR owing to the fact
that the trigger for processing in the old app was the prison receiving
the request email and that cancelation was handled by a manual process
of marrying up the cancelation emails with the request emails.