Enable visit editing from confirmation page
This adds a 'review_step' parameter that will override the default
behaviour of the StepsProcessor and be shown instead of the first
incomplete step.

In order to distinguish incomplete from not-yet-filled steps (necessary
to avoid showing validation errors on the first try), the hidden fields
for later steps are only shown when reviewing.

The option selectors for visit slots are returned to using Rails's
FormBuilder methods, reverting the change made in a2ca599, as this
ignored existing selections. As the correct CSS class is applied, the
JavaScript enhanced selector continues to work.

To implement the edit, there is a form with hidden fields for each step
on the confirmation page. As there are three editable steps, this means
that the fields are replicated three times. This is slightly
unfortunate, but works well.