Add Depersonalizer to scrub old records
In order to meet the agreed privacy requirements, we want to eliminate
personal data from the database after a month.

Personal data is defined for these purposes as:

* Prisoner Name
* Prisoner DOB
* Prisoner Number
* Prisoner Location
* Visitor Name
* Visitor DOB
* Visitor Phone No.
* Visitor Email
* Visitor IP Address

Of these, the application does not record IP addresses. Location is
required for metrics, but ceases to be personal when all the other
prisoner information has been scrubbed.

To retain information for metrics and so that we can continue to rely on
database constraints, we do this by replacing the sensitive data with
'REMOVED' or 1 January in the year 1 CE.  (There is, of course, no year
zero in our calendar.)