Raise exception if recurring slot is invalid
At present, this will only raise at runtime, but it moves us toward
being able to verify the prison data in advance.

Using exceptions is consistent with how other data types used in slot
configuration - i.e. Date - work. In future, if we add a front-end to
edit this data, we may not want it to raise exceptions. But at that
point we won't want it in the same format in any case.