Disable query caching for healthchecks
By default, `ActiveRecord::QueryCache` will try to connect to the
database before anything else happens to prepare cached queries, which
breaks for the healthcheck. We don't need the database in `ping.json`,
and `healthcheck.json` should be able to control when we connect to the
database (so it can report on the outcome), so connecting before either
isn't worth it.

In the old app, we got around this problem by completely disabling
query-caching, but that seems like a bad idea for a database-backed app.