Fix English-specific assumptions in slot picker JS
Specifically:

* am/pm are locale-dependent
* the abbreviated form of a month is not always the first three letters
* hour and minute are locale-dependent
* pluralisation does not always mean appending the letter s

It retains other unsatisfactory assumptions regarding spacing, ordering,
and left/right branching, but it lets us show the slot picker in Welsh.

Ideally, the JavaScript would not be responsible for building these
strings but would pick them up from the DOM. However, that is a fight
for another day.