Localise template header
This is a bit hacky: the template uses what it calls the
`proposition_title`, which is taken from `Rails.configuration`. This
doesn't allow for localisation per-request, and changing the
configuration on each action is unlikely to be thread-safe.

However, all configuration is taken from a method in `MojHelper`, and by
defining our own version of this we can replace the existing behaviour
and fetch the application title from `I18n.t`.