Update seed file format
As we have now synchronised our seed data, we can allow them to diverge.
We need them to diverge in order to support localisation of prison names
and addresses, and this is a good time to:

* Make the address a string, instead of joining lines
* Make the keys match those used internally

This means that when we add the translations, the translated addresses
can match the default addresses without having to do some custom
processing.