Don't blow away the database on deploy
As we are using the SQL schema format, `rake db:schema:load` will no
longer work. In its place, `db:setup` will work regardless of the schema
format.

On deploy, we should migrate the database, not blow it away.
Furthermore, as seeding is idempotent, there is no need to migrate
without seeding. We want to seed on every deployment in order to ensure
that the prisons are up to date and reflect the data on disk.