Redirect mails to maildrop based on ENV variables
Currently, staging sends real emails.  This changes the seeder such that
it writes a `maildrop.dsd.io` address for each prison if the
`SMTP_HOSTNAME` environment variable is set to `maildrop.dsd.io`.  I
decided to do it this way as it prevents us from having to change the
smoketests to account for different environments. This would have been
the case with any of the other mail interceptor gems.