Remove the broken scenic gem
We thought that Thoughtbot's scenic gem would help us deal with the
issue of `structure.sql` or `schema.rb` being inexplicably reordered on
different developers' machines.  However, after we migrated to scenic,
we discoverd that this version breaks if you try to run `rake
db:drop db:setup` or `rake db:drop db:create db:schema:load`. It loses
the connection to the database when it hits the first `create_view`
declaration.

I have reproduced the error on a newly commissioned Rails app and have
raise the issue with Thoughtbot:

https://github.com/thoughtbot/scenic/issues/129

For now, the best alternative is to return to the `structure.sql`
version we were using before.  This will still have the ordering
problem, but it works.