Replace logstasher with lograge
I decided to move logging into the controller, although the rationale for this is arguable.

The other solution would be to store additional data in something like request_store, but right now the complexity did not seem justified.

Again testing the logging of visit id seems excessive, so I removed rather than re-implementing.