Keep migrations from breaking on RDS
For some reason RDS does not like the `distribution` SQL. As yet, we have
been unable to ascertain the reason.  In order to allow other db
changes, including new prison seeds, to work without manual
intervention, I am removing the failing migraitons and the code they
support. We will revisit this sometime in the near future.