Accept visit while rejecting visitors
PVB1 allowed prisons to accept a visit while rejecting some of the
visitors.  This restores that functionality.  If all visitors are
rejected, the visit will be automatically rejected.  If some of the
visitors are rejected, the reasons will be included in the acceptance
email. At this time it is possible to accept at visit, but reject the
requesting visitor. This may need revisiting at a later date.