Fix bad validations when no slot is selected
After changing the flags to allow for banning or noting unlisted
visitors separately, the application would throw a 500 error when banned
or not_on_list were selected, but a slot was not selected. This was
because `_inquiry` does not gracefully handle blank values.

In order to prevent this, I set the default for `selection` to none.
Stupidly, I then added that to the global used to check that a slot had
been selected.  This meant that every record was coming back valid to
begin with, even if a time slot had not been selected.  It also meant
that the model was always requiring a reference number, even if there
was not actually a slot.

This change fixes the problem and *should* resolve the issue being seen
on some requests whereby they *appear* to reload the page without
showing any errors (they are expecting a reference number, even though
the selection is 'none').