Timeout after 2s when using the Sendgrid API
Use `Timeout.timeout` around the Sendgrid client.

Currently SendgridToolkit uses HTTP Party without a timeout so if the Sendgrid
API becomes unresponsive booking a visit becomes unresponsive.

Ideally the timeout would be set on the HTTP client, however it is not possible
to set it by design. We should revisit this issue as the fix is not a good term
solution.