Refactor initializer, connection pool and client
Improvements following a conversation with @mloughran

The initializer was passing in a hard to understand hash as configuration to the
pool which was a smell.

Refactored the code to inline the pool concept into `SendgridApi` to help with
the configuration problem.

Other improvements include drying up the rescues in `SendgridApi` and
simplifying the client to use a single `request` method instead of having a
method for `get` and `post` since they are so similar.

Also `StandardError` is being rescued and reported to Sentry to prevent the
service becoming unavailable if any problems arise in the sendgrid validation
code path.