Fix memoization issue with shared examples
In order to deterimine the origin of a counting error, I needed to make
sure the db view was partitioning the data correctly. This led me to
discover that the definitions for the overdue visits were being memoized
and were unable to make more than one visit per call per spec. To fix
this, I changed them from `let` declarations to methods.

This also includes the change showing that the additional overdue
bookings are correctly assigned to Luna.