Add essential validation to create booking API
As hinted in a comment, this method has complex validation requirements which should be handled in an object. However I have opted not to do so for now, but rather to wait until the inevitable refactoring when steps are removed.

This code at least ensures that bookings cannot be created that would fail the existing frontend validation.