'/status/:id' to catch pvb1 email url
Implements the pvb1 path that it is still being used. Users got a link on their
email in pvb1 that took them to a page with information about a visit.

It renders a page similar to the not found page but with a message that is more
helpful than just page not found.

Ommited the Welsh translation since it shoud always default to the english
locale.