Ensure weekly rejection percentages are correct
Previosly, the SQL used for the calculations would report the same
percentage for every type of rejection including the total. This was
because the date selection logic was in the wrong place.

This also removes the class and the specs for rejections done by
calendar date.  This was not being used anywhere and required additional
tuning to avoid the same problems.  If it is required, we can revisit
the SQL and migrations at a later date.