Use locale in the API from the headers
Uses the locale from the header in the API instead of using a redundant locale
parameter.

The specs uncovered some missing welsh localisation, I'll make a new PR for
this issue.

Also as discussed I've refactored out the gem we were using to parse the accept
language header as we don't use it's functionality.