Handle unknown formats in request a booking
Requesting a random format will cause Rails to try to search for a template with
a matching file extension. Since the file does not exist it raises a 500.

Adding `respond_to` in an action will only match on the formats specified by the
block. If a request is of an unknown format it raises an
`ActionController::UnknownFormat` which gets translated with a request with an
error code 406.

While this works for these endpoints it doesn't catch all endpoints, for that
there are 2 possible approaches:
  - one is to use the `responders` gem that was extracted from Rails 3 and use
  it on on the ApplicationController
  - change the routes so that they match on the format

I'd leave investigating this solutions until we deploy the public app and cleanup
pvb2.