Render the 404 template when responding with a 406
Otherwise without this the app will respond with a 406 status code and a blank
body.

I think that the 404 copy is the best for the users, for example if they request
'en/request.pdf' they'll get back a page saying that the page does not exist. I
think this is better than creating a new error page for 406 that says that the
format was not acceptable as it will probably make no sense to most of our
users.