Workaround for rake / rspec bug
The feature spec that checked the behaviour of visiting a not permitted locale
was failing when using rake due to a bug. The error appears to be raised outside
the example so the expecation fails. However the spec works correctly when
running it with rspec.

Implemented a workaround by using the url generators, these use the routes so
they raise an error when attempting to generate a url for a route that does not
exist.