Fix ZendeskErrors due to stricter email validation
Zendesk has stricter email validations than just trying to parse the email
(a recent change for pvb-public). This re-introduces email checker, but without
using Sendgrid for validations.

I spent a good time trying to extend EmailChecker to be better at detecting
formatting errors, or only formatting errors (without MX checks or Sendgrid) but
it became a rabbit hole in terms of design and complexity of validating properly
the format of an email address and domain part.

At the end I decided to stick with the current formatting checks and MX
validation as is a more incremental change. In the future we could look at
improving the validation by validating the domain format.