Fix no reply getting set to invalid value
With the recent change to configure the mailers in the initializer it meant that
the NoReply concern was being lodaded before the translations were available.
The result of this was the the 'from' address was set to a String from I18n
saying there was no translation available.

This branch fixes this by setting the 'from' field at runtime instead.