Set the puma threads max to be the db pool size
Fixes an issue where Puma can accept a request but error out because the DB pool
times out. This happened a couple of months ago when the Sendgrid API started
becoming unresponsive and we didn't have a pool of Sendgrid clients.

The fix prevents the puma and DB pools be out of sync by using the same source
for the max number of threads

This is similar to how Rails 5 will be configured by default with Puma [1].
Although with Rails 5 Puma can be configured via the command line with `rails s`
so we can revisit this when we upgrade.

[1] https://blog.heroku.com/archives/2016/5/2/container_ready_rails_5