Redirect when successfully processing a visit
Create a new page to render the result of processing a visit.

This will be needed to link to a cancellation form for staff and to render the
results of the cancellation.

This new page is temporary, ultimately we'll want to redirect to the new visit
show page from the dashboards, but only when we've released the dashboards.