Override the visitor cancel link in prison mailer
We think that prisons are using the cancel link in the copied visitor email that
they get. We are changing the value of the link so that instead of cancelling
the visit as a visitor, they'll cancel it as a prison.

This way will also let us see if prisons are using the link by looking into the
analytics.