Format Sidekiq logs in logstash format
Logstash only receives http app logs which means that we have no visibility on
how the asynchronous jobs have performed when debugging issues.

It works by broadcasting the Sidekiq logs to the logstash log file and merging
event data about the processing from the Sidekiq logs and ActiveJob
instrumentation. This is necessary because Sidekiq and ActiveJob do not log
enough data by themselves. This approach is similar to the Logstasher gem that
currently formats http requests for logstasher.