Fix acceptances being rejected in some scenarios
When the form had a selected banned or not on list visitor but the visitor
banned / not on contact list checkboxes were unchecked and the visit was
accepted the visit was process as rejected because there were no available
adults.

This happened because the frontend sends the visitor ids even though when the
checkbox is deselected and the backend only took into account the not allowed
visitor ids.

This commit fixes this by taking into account the deselected checkboxes and
ignoring any visitor ids that are sent. Ideally we'd also get the front end to
not send any visitor ids if the checkboxes are deselected.