Send Zendesk feedback to public or staff inboxes
The legacy inbox in Zendesk is set to match tickets that are from prison-visits
service. The new staff inbox matches tickets that have the 'staff.prison.visits'
tag.

Ideally both inboxes would be configured similarly by matching on service and a
tag. This is no possible because the legacy public tickets are not tagged. We
could look at tagging all past and new tickets and make the configuration
consistent.