Refactor SignonIdentity & current estate concerns
This avoids SignonIdentity being mutable, making it easier to reason about.

The current estate is stored simply as an id in the user's session, and should be easy to move into the URL path in future if desired without making changes to SignonIdentity.

This commit also moves all responsibility for available estates back into SignonIdentity.

There is no need to validate the current estate, because a default estate is always available via SignonIdentity (if there's no estate, you can't log in).