Update Pentonville visits
"We do not book social visits on thursday am " - we have Thursday 09:15 - 11:15 in the visits calendar for Pentonville. Is this not right? No- we have no social visits on Thursday, Friday and Sunday mornings. We also have no social visits in the afternoon on the first Tuesday of every month.
 
"Family day" - if you have family days in future, we can block these days out on your visits calendar. Just let us know 28 days in advance. Next family day is the 11th December. We will send you the 2017 calendar once we have finalised it
 
"you can only have 3 adults and 3 kids a visit. anyone over the age of 10 is concidered as an adult" - we can amend the adult age for Pentonville to 11+ if this would be helpful? This would stop people requesting visits with more than 3 people under/over 10 years old. Yes please that would be helpful