Verify past prisoner details in Nomis
Reads in the Nomis API configuration and the number of days of most recent
prisoners to check and prints out json entry for each prisoner check saying if
it was found or had and error.

By reading the Nomis API configuration we should be able to test in production
before configuring the server.