Retry Nomis API failed requests
This should solve EOFError caused by terminated persistent connections by the
server and other flakey errors that we are getting from the API.

Marking the request as idempotent (if possible) is what Excon recommends for
solving EOFError's. Internally idemptoent requests that fail have the connection
reset.

The retry limit is set to 2, the minimum, which will reset stale connections
and avoid hitting the rate limiting.