Staff slot availability (#683)
* Decouple the public slot from prisoner availability

* Add environment variables for slot availability

* Rename slot_availability to api_slot_availability

* Fix typo in spec

* Add slot checks to the staff nomis checker

* Update the api integration for prison slots

The api path and response body has been updated since we originally implemented
it.

It also wraps the response with an object to keep it consistent prisoner
availability.

* Make concrete slots comparable

* Prison slot validator for staff

Similar to the prisoner availability validator but instead of checking dates it
checks slots.

* Display slot availability when processing a visit

* Add slot unknown warning copy and date box styling

* Correct the slots times from the Nomis API

Nomis doesn't allow overlapping times so if there are 2 slots one after the
other one one of them is off by 1 minute. This means that this slots won't match
the selected slots by the visitors if they aren't corrected.

* Track errors of individual nomis api calls

We track the total number of api calls that fail, however we don't know which
ones are the ones failing.

* Add script to check the prison availability

Checks the prison availability on requested visits for a given prison.