Stop calculating unused distribution metrics
Removes the 99, 75, 25 percentiles which are unused.

I've tried removing the views and inlining the query but came up with 2 issues.

First one was that by using `ActiveRecord.connection.execute` it didn't decode
the array value in the result.

Second one was that we filter the query by prison and week, this is easy to do
with views by using `where` but becomes more complex when using raw SQL queries.