Add extra context to Sentry for bad nomis response
Looks like the detailed prisoner availability api from time to time it responds
with a bad response that breaks the code. The problem is that by the time we
investigate the request that broke the responses look normal.

This adds the response to the sentry context if it detects that the response is
non standard.