Replace timebox with a soft request deadline (#791)
* Custom Excon Idempotent middleware

The Idempotent middleware retries timeouts, we don't want this as we want to
return default responses so we don't break the SLA that public expects.

* Deadline Excon middleware

Introduces the concept of deadlines within Excon so that we can rely on network
timeouts instead of the timebox timeouts.

Takes a future time as an argument and sets the read / write timeouts based on
that future time. If the time has already elapsed it raises an error.

* Set middlewares only on the nomis connection

No need to change the Excon defaults to use our middlewares.

* Raise custom error with timeout or deadline errors

Reraise timeout or deadline errors as custom error so that we can customise
fallback responses in public.

* Set request deadline and delete timebox

Sets an overall request deadline for api requests and deletes the old timebox.