Fix excon instrumentation (#802)
* Bump pvb-instrumentation with retries fix

* Fix excon instrumentation

Excon's middleware stack doesn't instrument the time taken to read the response
from the socket due to the position of the Instrumentor middleware in the stack.

This PR moves the Instrumentation to the start of the stack and also changes the
payload of the instrumented event to not expect the response object to already
be set. Otherwise ActiveSuport::Instrumentation raises a nil error.