Update StaffResponse/BookingResponder interaction
This change will let us recover from errors when visits get booked to Nomis.
There is no change in behaviour, just a refactor.

The previous interaction relied that StaffReponse would run the validations, and
it it was valid then BookingResponder would execute the action and always succeed.
This will no longer be the case when booking to nomis since the api might fail
on the execution path.

The main change is to introduce a BookingResponse object that represents the
result of executing the action (accepting, rejecting, etc). This object will be
used to abort the transaction, not send emails and render errors back to the
user when booking to nomis.