Allow visits to be saved to Nomis in processor
Updates the BookingRequestProcessor to save visits to Nomis if configured to do
so.

The visit is created inside the database transaction to try to minimise the
chances of errors leaving Nomis and PVB inconsistent.

For example if the api is called was to be called outside the transaction then
we could end up with the visit persisted to nomis and the transaction failing,
or if it was called afterwards then pvb could end up with the visit booked but
not in nomis.