Merge the prison visits show pages
Until this change we had 2 pages to display a visit, one for visits on
'requested' state (process_visit) and another for the other states (show).

This was confusing as it was possible to end up in the show version of the page
for requested visits by modifying the url. That version of the page was very
confusing as it looked very similar to the other one so it was not obvious that
it was different page.

Now we have a single page that renders a different version depending on the
state of the visit.

It keeps the old routes around that don't have the locale in the url for
backwards compatibility, these can be deleted in a few days when staff don't
have open pages with the old urls.