Refactor env variables config parsing
Prevents having to define `nil` configuration in `test` environment.

This was a problem because when we had a new environment variable for
configuration that we didn't want to leak into the tests we had to remember
to explicitly set it in `config/environments/test.rb` which was easy to forget.