Fix link directory setup
The setup in the initializer was setting the publice service as a URI Template
value, however the implementation of LinkDirectory expects it to be url.

The consequences of this was that links to the public visit page in the emails
had a double '/' in the path ie http://example.com//en/visits/123, and urls
generated for the public service through `LinkDirectory#public_service` would
have had the path duplicated.