Fix email validation bug
The issue here is that older browsers are not compatible with the newer
HMTL input type "email", which alerts users when they have entered an
invalid email address (e.g. "user@example..com") when using the feedback
submission form.  This means that in older browsers form submissions
are allowed to go ahead regardless of an invalid email format, but then
result in an error in Sentry.  Therefore the #domain_dot_error method
has been extended to check for two or more consecutive dots in the
domain; if found an error is raised.