Avoid a memory leak in Excon
Current version of Excon has a memory leak when threads using an Excon
connection get destroyed.

Applies same fix as
https://github.com/ministryofjustice/prison-visits-public/pull/346

The memory leak in pvb2 is slower because Puma is configured to use less threads
so they probably get recycled less often.

Also sets a constant number of threads for Puma in production as we do in
pvb-public.