Add 'other' rejection reason checkbox
This commit will implement the addition of an 'other' rejection reason
for staff to use when processing visit requests.  Metrics have shown a
high number of rejection reasons as 'slot unavailable', but we now
understand that staff are using this rejection reason in lieu of an
option that actually matches their case.  The 'other' checkbox also has
a text field for staff to write down their reason for this rejection.
The aim is to monitor the 'other' rejection reasons and ensure that we
add checkboxes, where necessary, to cover a wider range of rejection
reasons that staff use on a regular basis.