Not hardcode confirmed bookings metrics date range
It is now possible to generate an export email with a custom range from the
console.

The wrapper class for the weekly prison counter view was hardcoded to pull
always the last 12 weeks of data regardless of what the exporter object was
configured to export.

It also had a bug if used within the first 3 months because the year was
hardcoded.