Fix weekly metrics new year bug
The export was missing data due to the query exluding data because of how the
filtering was done.

Because the query doesn't work with SQL date primitives working with year
boundaries is harder. Now if the export goes through a year boundary it will
issue 2 queries.

This won't work if the range spans 3 different years but I don't think this will
be a problem in practice.