Remove Virtus
Removes Virtus and uses the new upcoming feature in Rails 5.2 that supports
typed attributes.

The main difference between Virtus and Rails approach is that Virtus are not
really types, it uses coercers so the resulting type can be different. In Rails
you specify the final type.

Another difference is that Rails doesn't support typed arrays, what in Virtus
can be written as `Array[Date]` for example. The way I ported these is by
creating specific classes for each typed array. This is not ideal because there
is a few very similar classes. This is something that can be improved when the
Rails API evolves to support (ActiveRecord when used with postgres supports
array types but it hasn't been ported to ActiveModel)

The final main difference is that Virtus allowed undeclared fields to be used
while ActiveModel doesn't.

Fix slotavailability

Fix offender availability

Add undocumented fields

Add missing API fields

Use the prisoner number type on Prisoner