Handle missing attributes
Following an upgrade to Ruby 5.2 we had experienced an outage due to the
fact that we hadn't explicitly declared all the attributes expected to
be returned from the NOMIS API call.  This PR aims to allow us to
handle this more defensively/gracefully so the user can continue to
use the service, whilst we are sent an alert so we can make the
necessary changes.

Please see trello card:
https://trello.com/c/01fIsgcU/1351-handle-unexpected-attributes-in-nomis-api-responses-more-defensively
for more details.