Remove Sheppey Cluster group
As part of this Trello card:
https://trello.com/c/94sDa4Nd/1379-clean-up-team-staff-access-to-pvb-via-sign-on
we have made contact with various admins within the groups we have under
Sign On.  As a result we have been informed that the Sheppey Cluster is
no longer an entity, and all three prisons are back to working independently.
We have removed the Sheppey Cluster group from the estates file so that
staff will go back to only being able to access PVB for the prison they
work in.

There is also the finder_slug which references the cluster, but this is
related to Prison Finder; so until that has been updated we cannot remove
this.